Corrected the getReading implementation. Right now works only for Accelerometer

+++ MobileClients/Android/nervousnetLIB/src/ch/ethz/coss/nervousnet/lib/BatteryReading.java
@@ -156,7 +156,7 @@ public class BatteryReading extends SensorReading {
 	/**
 	 * @return the volt
 	 */
-	public long getVolt() {
+	public int getVolt() {
 		return volt;
 	}
